Pass builds: SHARED Model LifetimeStats move from APIFields -> models left stale imports
